+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/PrometheusHttpListenerOptions.cs
@@ -94,6 +94,12 @@ public class PrometheusHttpListenerOptions
     }
 
     /// <summary>
+    /// Gets or sets a value indicating whether to include a <c>target_info</c> metric in the scrape response.
+    /// Default value: <see langword="true"/>.
+    /// </summary>
+    public bool TargetInfoEnabled { get; set; } = true;
+
+    /// <summary>
     /// Gets or sets an optional callback to apply custom configuration for the
     /// <see cref="System.Net.HttpListener"/> instance used by the exporter.
     /// </summary>