+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/CHANGELOG.md
@@ -26,6 +26,14 @@ Notes](../../RELEASENOTES.md).
   disable scope labels in Prometheus metrics. Defaults to `true`.
   ([#7436](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7436))
 
+* Added support for the `dots` and `values` Prometheus UTF-8 name escaping
+  schemes when negotiated via the `Accept` header.
+  ([#7439](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7439))
+
+* Add `PrometheusHttpListenerOptions.TargetInfoEnabled` property to enable or
+  disable the `target_info` metric in Prometheus metrics. Defaults to `true`.
+  ([#7438](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7438))
+
 * Added the `PrometheusHttpListenerOptions.ConfigureHttpListener` option.
   This can be used as a replacement for the removed `UriPrefixes` option.
   ([#7448](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7448))